+++ tests/MiniZinc.ClientTests/IntegrationTests.fs
@@ -4,293 +4,277 @@ open MiniZinc
 open Xunit
 open System.IO
 
-module ``Integration Tests`` =
+type ``Integration Tests``(fixture: ClientFixture) =
     
-    
-    [<Fact>]
-    let ``test aaaa`` () =
-//         let mzn = """forall(i in 1..n-1)(
-// 		if true
-//         then
-//         sum(j in 1..t)(1)
-//         else
-//         1
-//         endif
-// );
-// """
-        let mzn = """if true then sum(j in 1..t)(1) else true endif;"""
-        let model =
-            Parse.parseString Parsers.if_else_expr mzn
-            |> Result.get
-        ()
-    
-   
+    let client = fixture.Client
+            
     let test (name: string) =
         let suite = TestSuite.load name
         let model = TestSuite.parseModel suite
+        let result = client.SolveAndWait(model)
         ()
-
+        
+    interface IClassFixture<ClientFixture>
 
     [<Fact>]
-    let ``test 2DPacking`` () =
+    member this. ``test 2DPacking`` () =
         test "2DPacking.mzn"
 
-
     [<Fact>]
-    let ``test alpha`` () =
+    member this. ``test alpha`` () =
         test "alpha.mzn"
 
-
     [<Fact>]
-    let ``test battleships10`` () =
+    member this. ``test battleships10`` () =
         test "battleships10.mzn"
 
 
     [<Fact>]
-    let ``test battleships 1`` () =
+    member this. ``test battleships 1`` () =
         test "battleships_1.mzn"
 
 
     [<Fact>]
-    let ``test battleships 2`` () =
+    member this. ``test battleships 2`` () =
         test "battleships_2.mzn"
 
 
     [<Fact>]
-    let ``test battleships 3`` () =
+    member this. ``test battleships 3`` () =
         test "battleships_3.mzn"
 
 
     [<Fact>]
-    let ``test battleships 4`` () =
+    member this. ``test battleships 4`` () =
         test "battleships_4.mzn"
 
 
     [<Fact>]
-    let ``test battleships 5`` () =
+    member this. ``test battleships 5`` () =
         test "battleships_5.mzn"
 
 
     [<Fact>]
-    let ``test battleships 7`` () =
+    member this. ``test battleships 7`` () =
         test "battleships_7.mzn"
 
 
     [<Fact>]
-    let ``test battleships 9`` () =
+    member this. ``test battleships 9`` () =
         test "battleships_9.mzn"
 
 
     [<Fact>]
-    let ``test blocksworld instance 1`` () =
+    member this. ``test blocksworld instance 1`` () =
         test "blocksworld_instance_1.mzn"
 
 
     [<Fact>]
-    let ``test blocksworld instance 2`` () =
+    member this. ``test blocksworld instance 2`` () =
         test "blocksworld_instance_2.mzn"
 
 
     [<Fact>]
-    let ``test cutstock`` () =
+    member this. ``test cutstock`` () =
         test "cutstock.mzn"
 
 
     [<Fact>]
-    let ``test eq20`` () =
+    member this. ``test eq20`` () =
         test "eq20.mzn"
 
 
     [<Fact>]
-    let ``test factory planning instance`` () =
+    member this. ``test factory planning instance`` () =
         test "factory_planning_instance.mzn"
 
 
     [<Fact>]
-    let ``test golomb`` () =
+    member this. ``test golomb`` () =
         test "golomb.mzn"
 
 
     [<Fact>]
-    let ``test halfreif`` () =
+    member this. ``test halfreif`` () =
         test "halfreif.mzn"
 
 
     [<Fact>]
-    let ``test jobshop2x2`` () =
+    member this. ``test jobshop2x2`` () =
         test "jobshop2x2.mzn"
 
 
     [<Fact>]
-    let ``test knights`` () =
+    member this. ``test knights`` () =
         test "knights.mzn"
 
 
     [<Fact>]
-    let ``test langford`` () =
+    member this. ``test langford`` () =
         test "langford.mzn"
 
 
     [<Fact>]
-    let ``test langford2`` () =
+    member this. ``test langford2`` () =
         test "langford2.mzn"
 
 
     [<Fact>]
-    let ``test latin squares fd`` () =
+    member this. ``test latin squares fd`` () =
         test "latin_squares_fd.mzn"
 
 
     [<Fact>]
-    let ``test magicsq 3`` () =
+    member this. ``test magicsq 3`` () =
         test "magicsq_3.mzn"
 
 
     [<Fact>]
-    let ``test magicsq 4`` () =
+    member this. ``test magicsq 4`` () =
         test "magicsq_4.mzn"
 
 
     [<Fact>]
-    let ``test magicsq 5`` () =
+    member this. ``test magicsq 5`` () =
         test "magicsq_5.mzn"
 
 
     [<Fact>]
-    let ``test multidimknapsack simple`` () =
+    member this. ``test multidimknapsack simple`` () =
         test "multidimknapsack_simple.mzn"
 
 
     [<Fact>]
-    let ``test oss`` () =
+    member this. ``test oss`` () =
         test "oss.mzn"
 
 
-    [<Fact>]
-    let ``test packing`` () =
+    [<Fact(Skip="Slow")>]
+    member this. ``test packing`` () =
         test "packing.mzn"
 
 
     [<Fact>]
-    let ``test perfsq`` () =
+    member this. ``test perfsq`` () =
         test "perfsq.mzn"
 
 
     [<Fact>]
-    let ``test perfsq2`` () =
+    member this. ``test perfsq2`` () =
         test "perfsq2.mzn"
 
 
-    [<Fact>]
-    let ``test photo`` () =
+    [<Fact(Skip="Slow")>]
+    member this. ``test photo`` () =
         test "photo.mzn"
 
 
-    [<Fact>]
-    let ``test product fd`` () =
+    [<Fact(Skip="Slow")>]
+    member this. ``test product fd`` () =
         test "product_fd.mzn"
 
 
     [<Fact>]
-    let ``test product lp`` () =
+    member this. ``test product lp`` () =
         test "product_lp.mzn"
 
 
     [<Fact>]
-    let ``test quasigroup qg5`` () =
+    member this. ``test quasigroup qg5`` () =
         test "quasigroup_qg5.mzn"
 
 
     [<Fact>]
-    let ``test queen cp2`` () =
+    member this. ``test queen cp2`` () =
         test "queen_cp2.mzn"
 
 
     [<Fact>]
-    let ``test queen ip`` () =
+    member this. ``test queen ip`` () =
         test "queen_ip.mzn"
 
 
     [<Fact>]
-    let ``test radiation`` () =
+    member this. ``test radiation`` () =
         test "radiation.mzn"
 
 
     [<Fact>]
-    let ``test simple sat`` () =
+    member this. ``test simple sat`` () =
         test "simple_sat.mzn"
 
 
     [<Fact>]
-    let ``test singHoist2`` () =
+    member this. ``test singHoist2`` () =
         test "singHoist2.mzn"
 
 
     [<Fact>]
-    let ``test steiner-triples`` () =
+    member this. ``test steiner-triples`` () =
         test "steiner-triples.mzn"
 
 
     [<Fact>]
-    let ``test sudoku`` () =
+    member this. ``test sudoku`` () =
         test "sudoku.mzn"
 
 
     [<Fact>]
-    let ``test template design`` () =
+    member this. ``test template design`` () =
         test "template_design.mzn"
 
 
     [<Fact>]
-    let ``test tenpenki 1`` () =
+    member this. ``test tenpenki 1`` () =
         test "tenpenki_1.mzn"
 
 
-    [<Fact>]
-    let ``test tenpenki 2`` () =
+    [<Fact(Skip="Slow")>]
+    member this. ``test tenpenki 2`` () =
         test "tenpenki_2.mzn"
 
-
+    
     [<Fact>]
-    let ``test tenpenki 3`` () =
+    member this. ``test tenpenki 3`` () =
         test "tenpenki_3.mzn"
 
 
     [<Fact>]
-    let ``test tenpenki 4`` () =
+    member this. ``test tenpenki 4`` () =
         test "tenpenki_4.mzn"
 
 
     [<Fact>]
-    let ``test tenpenki 5`` () =
+    member this. ``test tenpenki 5`` () =
         test "tenpenki_5.mzn"
 
 
     [<Fact>]
-    let ``test tenpenki 6`` () =
+    member this. ``test tenpenki 6`` () =
         test "tenpenki_6.mzn"
 
 
     [<Fact>]
-    let ``test timetabling`` () =
+    member this. ``test timetabling`` () =
         test "timetabling.mzn"
 
 
     [<Fact>]
-    let ``test trucking`` () =
+    member this. ``test trucking`` () =
         test "trucking.mzn"
 
 
     [<Fact>]
-    let ``test warehouses`` () =
+    member this. ``test warehouses`` () =
         test "warehouses.mzn"
 
 
     [<Fact>]
-    let ``test wolf goat cabbage`` () =
+    member this. ``test wolf goat cabbage`` () =
         test "wolf_goat_cabbage.mzn"
 
 
     [<Fact>]
-    let ``test zebra`` () =
+    member this. ``test zebra`` () =
         test "zebra.mzn"
+